requirements.txt
Flask==3.0.2
python-telegram-bot==13.15
requests==2.31.0
boto3>=1.26.79
pytest>=7.2.1
botocore~=1.34.72
openpyxl
python-dotenv==1.0.1
pillow==10.2.0
celery==5.3.0
redis==5.0.3
sentence-transformers
#pysqlite-binary
chromadb-pysqlite3
openai
onnxruntime
python-docx
langchain~=0.1.16
langchain-core
langchain-community
langchain-openai
PyPDF2~=3.0.1